RetryBackoffCalculatorTesting.kt: created with shared testing logic to keep test DAMP (https://abseil.io/resources/swe-book/html/ch12.html)
